+++ apps/backend/src/db/migrations/20240823115158_create_onboarding_questions_table.ts
@@ -1,0 +1,31 @@
+import { type Knex } from "knex";
+
+const TABLE_NAME = "onboarding_questions";
+
+const ColumnName = {
+	CREATED_AT: "created_at",
+	ID: "id",
+	LABEL: "label",
+	UPDATED_AT: "updated_at",
+} as const;
+
+async function up(knex: Knex): Promise<void> {
+	await knex.schema.createTable(TABLE_NAME, (table) => {
+		table.increments(ColumnName.ID).primary();
+		table.string(ColumnName.LABEL).notNullable().unique();
+		table
+			.dateTime(ColumnName.CREATED_AT)
+			.notNullable()
+			.defaultTo(knex.fn.now());
+		table
+			.dateTime(ColumnName.UPDATED_AT)
+			.notNullable()
+			.defaultTo(knex.fn.now());
+	});
+}
+
+async function down(knex: Knex): Promise<void> {
+	await knex.schema.dropTableIfExists(TABLE_NAME);
+}
+
+export { down, up };