+++ apps/frontend/src/libs/components/notification/notification.tsx
@@ -1,7 +1,7 @@
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 
-const Notification = () => {
+const Notification: React.FC = () => {
 	return <ToastContainer />;
 };