+++ apps/frontend/src/libs/hooks/use-app-form/use-app-form.hook.ts
@@ -5,7 +5,8 @@ import {
 	type FieldErrors,
 	type FieldValues,
 	type UseFormHandleSubmit,
-	UseFormProps,
+	type UseFormProps,
+	type UseFormSetError,
 	type ValidationMode,
 } from "react-hook-form";
 import { useForm } from "react-hook-form";
@@ -22,6 +23,7 @@ type ReturnValue<T extends FieldValues = FieldValues> = {
 	control: Control<T, null>;
 	errors: FieldErrors<T>;
 	handleSubmit: UseFormHandleSubmit<T>;
+	setError: UseFormSetError<T>;
 };
 
 const useAppForm = <T extends FieldValues = FieldValues>({
@@ -45,12 +47,14 @@ const useAppForm = <T extends FieldValues = FieldValues>({
 		control,
 		formState: { errors },
 		handleSubmit,
+		setError,
 	} = useForm<T>(parameters);
 
 	return {
 		control,
 		errors,
 		handleSubmit,
+		setError,
 	};
 };