+++ apps/frontend/src/modules/auth/slices/auth.ts
@@ -1,8 +1,9 @@
-import { signIn, signUp } from "./actions.js";
+import { getAuthenticatedUser, signIn, signUp } from "./actions.js";
 import { actions } from "./auth.slice.js";
 
 const allActions = {
 	...actions,
+	getAuthenticatedUser,
 	signIn,
 	signUp,
 };