+++ apps/frontend/src/pages/auth/libs/components/sign-in-form/sign-in-form.tsx
@@ -1,0 +1,60 @@
+import { Button, Input } from "~/libs/components/components.js";
+import { useAppForm, useCallback } from "~/libs/hooks/hooks.js";
+import {
+	type UserSignInRequestDto,
+	userSignInValidationSchema,
+} from "~/modules/users/users.js";
+
+import { DEFAULT_SIGN_IN_PAYLOAD } from "./libs/constants.js";
+import styles from "./styles.module.css";
+
+type Properties = {
+	onSubmit: (payload: UserSignInRequestDto) => void;
+};
+
+const SignInForm: React.FC<Properties> = ({ onSubmit }: Properties) => {
+	const { control, errors, handleSubmit } = useAppForm<UserSignInRequestDto>({
+		defaultValues: DEFAULT_SIGN_IN_PAYLOAD,
+		validationSchema: userSignInValidationSchema,
+	});
+
+	const handleFormSubmit = useCallback(
+		(event_: React.BaseSyntheticEvent): void => {
+			void handleSubmit(onSubmit)(event_);
+		},
+		[handleSubmit, onSubmit],
+	);
+
+	return (
+		<>
+			<form className={styles["form"]} onSubmit={handleFormSubmit}>
+				<Input
+					control={control}
+					errors={errors}
+					label="Email"
+					name="email"
+					placeholder="name@example.com"
+					type="email"
+				/>
+
+				<Input
+					control={control}
+					errors={errors}
+					label="Password"
+					name="password"
+					placeholder="*******"
+					type="password"
+				/>
+
+				<Button label="SIGN IN" type="submit" variant="dark" />
+			</form>
+
+			<div className={styles["circle-gradient1"]} />
+			<div className={styles["circle-gradient2"]} />
+			<div className={styles["circle-gradient3"]} />
+			<div className={styles["circle-gradient4"]} />
+		</>
+	);
+};
+
+export { SignInForm };