+++ apps/frontend/src/pages/auth/libs/components/sign-up-form/libs/constants.ts
@@ -1,0 +1,10 @@
+import { type UserSignUpFormDto } from "~/modules/users/users.js";
+
+const DEFAULT_SIGN_UP_PAYLOAD: UserSignUpFormDto = {
+	confirmPassword: "",
+	email: "",
+	name: "",
+	password: "",
+};
+
+export { DEFAULT_SIGN_UP_PAYLOAD };