+++ apps/mobile/src/libs/packages/toast-message/toast-message.ts
@@ -3,5 +3,4 @@ import { ToastMessage } from "./toast-message.package";
 const toastMessage = new ToastMessage();
 
 export { toastMessage };
-
 export { DEFAULT_ERROR_MESSAGE } from "./libs/constants/constants";