+++ apps/mobile/src/screens/auth/styles.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+import { BaseColor } from "~/libs/enums/enums";
+
+const styles = StyleSheet.create({
+	formContainer: {
+		backgroundColor: BaseColor.BG_WHITE,
+		borderRadius: 25,
+		width: "90%",
+	},
+});
+
+export { styles };