+++ packages/shared/src/index.ts
@@ -30,7 +30,10 @@ export {
 	type ValidationSchema,
 	type ValueOf,
 } from "./libs/types/types.js";
-export { AuthApiPath } from "./modules/auth/auth.js";
+export {
+	AuthApiPath,
+	ConfirmPasswordCustomValidation,
+} from "./modules/auth/auth.js";
 export {
 	type UserDto,
 	type UserGetAllResponseDto,
@@ -38,6 +41,7 @@ export {
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,
 	userSignInValidationSchema,
+	type UserSignUpFormDto,
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
 	userSignUpValidationSchema,